% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bb}
\alias{bb}
\alias{bb.sdr}
\title{Create a plot of a beachball}
\usage{
bb.sdr(strike, dip, rake, filename = NULL, load.results = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{strike,dip,rake}{numeric; the strike, dip, and rake of the moment tensor, in degrees}

\item{filename}{character; the name of the file to create}

\item{load.results}{logical; should the resuling file be loaded for plotting, etc?}

\item{verbose}{logical; should warnings and messages be given?}

\item{...}{additional parameters}
}
\value{
An object of class \code{\link{mtbb}}
}
\description{
Create a plot of a beachball
}
\note{
\code{\link{bb.sdr}} allows creation a beach ball without deviatoric components by specifying
the strike, dip, and rake of the earthquake.

\code{obspy-mopad} will rearrange the angles if they don't make sense!
}
\examples{
\dontrun{
#
# Create "beachball" representations with
#       ** Strike, Dip, Rake **
# of earthquake moment tensor
#
b1 <- bb.sdr(10,1,10) # succeeds
b2 <- bb.sdr(1,10,10) # values are rearranged internally
all.equal(b1, b2)

# change output
bb.sdr(10,1,10, filename='beachball') # writes to 'beachball.svg'

# Visualize:
bb <- bb.sdr(45,80,10)
plot(c(100, 250), c(300, 450), type = "n", xlab = "", ylab = "", asp=1)
rasterImage(b1[['image']], 100, 400, 150, 450)
rasterImage(bb[['image']], 170, 370, 220, 320)
}
}
\seealso{
\code{\link{mtbb-class}}
}

