% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pal.R
\name{color_pal}
\alias{color_pal}
\title{Beachball Color Palette}
\usage{
color_pal(n, type = "discrete", reverse = FALSE, levels = n)
}
\arguments{
\item{n}{number of palette colors}

\item{type}{palette type}

\item{reverse}{reverses palette order}

\item{levels}{number of colors to return}
}
\value{
A character vector of color hexcodes
}
\description{
Returns a color palette as a vector of RGB color hexcodes.
}
