% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_bg_color.R
\docType{data}
\name{d_bg_color}
\alias{d_bg_color}
\title{Dark Background Color}
\format{An object of class \code{character} of length 1.}
\usage{
d_bg_color
}
\description{
Defines the background color hex code for dark themes to use
}
\keyword{datasets}
