% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_legend.R
\docType{data}
\name{hide_legend}
\alias{hide_legend}
\title{Hide Legend}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 4.}
\usage{
hide_legend
}
\description{
Hides the legend of a ggplot visualization
}
\keyword{datasets}
