% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_x_gridlines.R
\docType{data}
\name{hide_x_gridlines}
\alias{hide_x_gridlines}
\title{Hide X Gridlines}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 2.}
\usage{
hide_x_gridlines
}
\description{
Hides gridlines parallel to the x axis in a ggplot visualization
}
\keyword{datasets}
