% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_y_gridlines.R
\docType{data}
\name{hide_y_gridlines}
\alias{hide_y_gridlines}
\title{Hide Y Gridlines}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 2.}
\usage{
hide_y_gridlines
}
\description{
Hides gridlines parallel to the y axis in a ggplot visualization
}
\keyword{datasets}
