% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mtbb-class}
\alias{mtbb}
\alias{mtbb-class}
\alias{points.mtbb}
\title{Methods for class \code{mtbb}}
\usage{
\method{points}{mtbb}(x, X, Y, M = rep(1, length(x)), M.max = max(M, na.rm =
  TRUE), bb.unit = "cm", add.M = FALSE, ...)
}
\arguments{
\item{x}{the \code{mtbb} object to plot}

\item{X,Y}{numeric; the x and y positions to plot the beachball at}

\item{M}{numeric; the magnitude of the earthquake}

\item{M.max}{numeric; the maximum magnitude to normalize the sizes by}

\item{bb.unit}{character; the \code{\link[grid]{unit}} to convert to}

\item{add.M}{logical; should \code{M} be shown with the beachball?}

\item{...}{additional parameters sent to \code{\link{text}} if \code{add.M=TRUE}}
}
\description{
Methods for class \code{mtbb}
}
\examples{
\dontrun{
b1 <- bb.sdr(45,80,70)
plot(-1:1, -1:1, type="b")
points(b1)
points(b1, X=0.1, add.M=TRUE)
}
}

