% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_day.R
\name{theme_day}
\alias{theme_day}
\title{Beachball Light Theme}
\usage{
theme_day(..., base_size = 14)
}
\arguments{
\item{...}{Other parameters to pass to theme_light}

\item{base_size}{The base size of elements}
}
\value{
A ggplot theme
}
\description{
A refined ggplot theme based on theme_light
}
