% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Beachball Map Theme}
\usage{
theme_map(..., base_size = 14, dark_theme = FALSE)
}
\arguments{
\item{...}{Other parameters to pass to theme_light}

\item{base_size}{The base size of elements}

\item{dark_theme}{Use theme_night for base theme}
}
\value{
A ggplot theme
}
\description{
A refined ggplot theme based on theme_light for creating maps/choropleths
}
