% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_night.R
\name{theme_night}
\alias{theme_night}
\title{Beachball Dark Theme}
\usage{
theme_night(..., base_size = 14)
}
\arguments{
\item{...}{Other parameters to pass to theme_light}

\item{base_size}{The base size of elements}
}
\value{
A ggplot theme
}
\description{
A refined ggplot theme based on theme_light with a dark background
}
